+++ server/impl/src/main/java/com/walmartlabs/concord/server/process/Payload.java
@@ -19,6 +19,7 @@ public class Payload {
     public static final HeaderKey<String> RESUME_EVENT_NAME = HeaderKey.register("_resumeEventName", String.class);
     public static final HeaderKey<ProjectDefinition> PROJECT_DEFINITION = HeaderKey.register("_projectDef", ProjectDefinition.class);
     public static final HeaderKey<Collection<String>> ACTIVE_PROFILES = HeaderKey.registerCollection("_activeProfiles");
+    public static final HeaderKey<Set<String>> PROCESS_TAGS = HeaderKey.registerSet("_processTags");
 
     public static final AttachmentKey WORKSPACE_ARCHIVE = AttachmentKey.register("archive");