+++ server/impl/src/main/java/com/walmartlabs/concord/server/process/pipelines/ProjectPipeline.java
@@ -30,6 +30,7 @@ public class ProjectPipeline extends Pipeline {
                 InventoryProcessor.class,
                 AttachmentStoringProcessor.class,
                 RequestDataMergingProcessor.class,
+                TagsExtractingProcessor.class,
                 PrivateKeyProcessor.class,
                 TemplateFilesProcessor.class,
                 TemplateScriptProcessor.class,