+++ server/impl/src/main/java/com/walmartlabs/concord/server/process/queue/ProcessQueueDao.java
@@ -1,21 +1,19 @@
 package com.walmartlabs.concord.server.process.queue;
 
 import com.walmartlabs.concord.common.db.AbstractDao;
+import com.walmartlabs.concord.common.db.PgUtils;
 import com.walmartlabs.concord.server.api.process.ProcessEntry;
 import com.walmartlabs.concord.server.api.process.ProcessKind;
 import com.walmartlabs.concord.server.api.process.ProcessStatus;
 import com.walmartlabs.concord.server.jooq.tables.records.ProcessQueueRecord;
-import org.jooq.Configuration;
-import org.jooq.DSLContext;
-import org.jooq.SelectWhereStep;
+import org.jooq.*;
 import org.jooq.exception.DataAccessException;
 import org.jooq.impl.DSL;
 
 import javax.inject.Inject;
 import javax.inject.Named;
 import java.sql.Timestamp;
-import java.util.List;
-import java.util.UUID;
+import java.util.*;
 
 import static com.walmartlabs.concord.server.jooq.tables.ProcessQueue.PROCESS_QUEUE;
 import static org.jooq.impl.DSL.currentTimestamp;
@@ -72,13 +70,27 @@ public class ProcessQueueDao extends AbstractDao {
     }
 
     public void update(UUID instanceId, ProcessStatus status) {
-        tx(tx -> update(tx, instanceId, status));
+        update(instanceId, status, Optional.empty());
+    }
+
+    public void update(UUID instanceId, ProcessStatus status, Optional<Set<String>> tags) {
+        tx(tx -> update(tx, instanceId, status, tags));
     }
 
     public void update(DSLContext tx, UUID instanceId, ProcessStatus status) {
-        int i = tx.update(PROCESS_QUEUE)
+        update(tx, instanceId, status, Optional.empty());
+    }
+
+    public void update(DSLContext tx, UUID instanceId, ProcessStatus status, Optional<Set<String>> tags) {
+        UpdateSetMoreStep<ProcessQueueRecord> q = tx.update(PROCESS_QUEUE)
                 .set(PROCESS_QUEUE.CURRENT_STATUS, status.toString())
-                .set(PROCESS_QUEUE.LAST_UPDATED_AT, currentTimestamp())
+                .set(PROCESS_QUEUE.LAST_UPDATED_AT, currentTimestamp());
+
+        if (tags.isPresent()) {
+            q.set(PROCESS_QUEUE.PROCESS_TAGS, toArray(tags.get()));
+        }
+
+        int i = q
                 .where(PROCESS_QUEUE.INSTANCE_ID.eq(instanceId))
                 .execute();
 
@@ -135,10 +147,10 @@ public class ProcessQueueDao extends AbstractDao {
     }
 
     public List<ProcessEntry> list() {
-        return list(null, DEFAULT_LIST_LIMIT);
+        return list(null, null, DEFAULT_LIST_LIMIT);
     }
 
-    public List<ProcessEntry> list(Timestamp beforeCreatedAt, int limit) {
+    public List<ProcessEntry> list(Timestamp beforeCreatedAt, Set<String> tags, int limit) {
         try (DSLContext tx = DSL.using(cfg)) {
             SelectWhereStep<ProcessQueueRecord> q = tx.selectFrom(PROCESS_QUEUE);
 
@@ -146,19 +158,42 @@ public class ProcessQueueDao extends AbstractDao {
                 q.where(PROCESS_QUEUE.CREATED_AT.lessThan(beforeCreatedAt));
             }
 
+            filterByTags(q, tags);
+
             return q.orderBy(PROCESS_QUEUE.CREATED_AT.desc())
                     .limit(limit)
                     .fetch(ProcessQueueDao::toEntry);
         }
     }
 
-    public List<ProcessEntry> list(UUID parentInstanceId) {
+    public List<ProcessEntry> list(UUID parentInstanceId, Set<String> tags) {
         try (DSLContext tx = DSL.using(cfg)) {
-            return tx.selectFrom(PROCESS_QUEUE)
-                    .where(PROCESS_QUEUE.PARENT_INSTANCE_ID.eq(parentInstanceId))
-                    .orderBy(PROCESS_QUEUE.CREATED_AT.desc())
+            SelectConditionStep<ProcessQueueRecord> q = tx.selectFrom(PROCESS_QUEUE)
+                    .where(PROCESS_QUEUE.PARENT_INSTANCE_ID.eq(parentInstanceId));
+
+            filterByTags(q, tags);
+
+            return q.orderBy(PROCESS_QUEUE.CREATED_AT.desc())
                     .fetch(ProcessQueueDao::toEntry);
         }
+    }
+
+    private Select<ProcessQueueRecord> filterByTags(SelectWhereStep<ProcessQueueRecord> q, Set<String> tags) {
+        if (tags == null || tags.isEmpty()) {
+            return q;
+        }
+
+        String[] as = tags.toArray(new String[tags.size()]);
+        return q.where(PgUtils.contains(PROCESS_QUEUE.PROCESS_TAGS, as));
+    }
+
+    private SelectConditionStep<ProcessQueueRecord> filterByTags(SelectConditionStep<ProcessQueueRecord> q, Set<String> tags) {
+        if (tags == null || tags.isEmpty()) {
+            return q;
+        }
+
+        String[] as = tags.toArray(new String[tags.size()]);
+        return q.and(PgUtils.contains(PROCESS_QUEUE.PROCESS_TAGS, as));
     }
 
     public boolean exists(UUID instanceId) {
@@ -177,6 +212,13 @@ public class ProcessQueueDao extends AbstractDao {
             kind = ProcessKind.DEFAULT;
         }
 
+        Set<String> tags = null;
+        String[] as = r.getProcessTags();
+        if (as != null && as.length > 0) {
+            tags = new HashSet<>(as.length);
+            Collections.addAll(tags, as);
+        }
+
         return new ProcessEntry(r.getInstanceId(),
                 kind,
                 r.getParentInstanceId(),
@@ -185,6 +227,15 @@ public class ProcessQueueDao extends AbstractDao {
                 r.getInitiator(),
                 r.getLastUpdatedAt(),
                 ProcessStatus.valueOf(r.getCurrentStatus()),
-                r.getLastAgentId());
+                r.getLastAgentId(),
+                tags);
+    }
+
+    private static String[] toArray(Set<String> s) {
+        if (s == null || s.isEmpty()) {
+            return new String[0];
+        }
+
+        return s.toArray(new String[s.size()]);
     }
 }